refactor(frontend): extract match draft helpers
